In-layer concurrency
This involves significant restructuring of CUDASimulation, and use of cudaStream_t in many (but not all) places.

A (incomplete) set of tests designed to detect concurrency are provided, through a new CUDASimulation CUDAConfig parameter (inLayerConcurrency)

Utility classes for timers are provided (CudaEvent and chrono::steady_clock based), and used to time RTCinit, init functions, step functions, exit functions and calls to simulate and per step timing.
These are programatically accessible through methods on the CUDASimulation object.
WDDM and cudaEvent based timers can be a little odd, it may ignore some time spent on the host. Further investigation required.

There is still room to make further improvements through streams (i.e. random init) and reduce the number of explicit and implicit synchronisation points required.
Pinned memory will also help in places. An issue will be created to highlight remaining things to do.

Adds some disabled tests to detect concurrency in the code executing after an agent condition or agent function. These cannot be fixed without pinned memory + a lot of refactoring, for minimal perf gain.

+++ include/flamegpu/gpu/CUDAEnsemble.h
@@ -77,9 +77,8 @@ class CUDAEnsemble {
     void setExitLog(const LoggingConfig &exitConfig);
     /**
      * Get the duration of the last call to simulate() in milliseconds. 
-     * With a resolution of around 0.5 microseconds (cudaEventElapsedtime)
      */
-    int64_t getEnsembleElapsedTime() const { return ensemble_elapsed_time; }
+    float getEnsembleElapsedTime() const { return ensemble_elapsed_time; }
     /**
      * Return the list of logs collected from the last call to simulate()
      */
@@ -117,7 +116,7 @@ class CUDAEnsemble {
     /**
      * Runtime of previous call to simulate() in milliseconds, initially 0
      */
-    int64_t ensemble_elapsed_time = 0;
+    float ensemble_elapsed_time = 0.f;
 };
 
 #endif  // INCLUDE_FLAMEGPU_GPU_CUDAENSEMBLE_H_
